consolidate cypress page tests into one file
