+++ src/applications/simple-forms/21-4138/tests/e2e/e2eHelpers.js
@@ -1,0 +1,7 @@
+import manifest from '../../manifest.json';
+
+const { rootUrl } = manifest;
+
+export const pagePathIsCorrect = pagePath => {
+  cy.location('pathname').should('eq', `${rootUrl}/${pagePath}`);
+};
